+++ CalendarEvents/src/main/java/me/nikl/calendarevents/Timing.java
@@ -71,13 +71,11 @@ class Timing {
 	 * Set up the relevant dates
 	 */
 	public void setUp(){
-		if(Main.debug){
-			Bukkit.getConsoleSender().sendMessage("days: " + days.toString());
-			Bukkit.getConsoleSender().sendMessage("dates: " + dates.toString());
-			Bukkit.getConsoleSender().sendMessage("monthlyDates: " + monthlyDates.toString());
-			Bukkit.getConsoleSender().sendMessage("yearlyDates: " + yearlyDates.toString());
-			Bukkit.getConsoleSender().sendMessage("times: " + times.toString());
-		}
+		Main.debug("days: " + days.toString());
+		Main.debug("dates: " + dates.toString());
+		Main.debug("monthlyDates: " + monthlyDates.toString());
+		Main.debug("yearlyDates: " + yearlyDates.toString());
+		Main.debug("times: " + times.toString());
 
 		// get references for current time (need to renew)
 		ZonedDateTime now = ZonedDateTime.now();
@@ -148,7 +146,7 @@ class Timing {
 				}
 			}
 		}
-		if(Main.debug)Bukkit.getConsoleSender().sendMessage("loaded " + relevantDates.size() + " days");
+		Main.debug("loaded " + relevantDates.size() + " days");
 		
 		for(ZonedDateTime date : relevantDates){
 			for(String timeString : times){
@@ -197,12 +195,12 @@ class Timing {
 				}
 			}
 			if(relevantMillis.isEmpty()){
-				Bukkit.getConsoleSender().sendMessage("[CalendarEvents] "+ ChatColor.RED+"All events with the label '" + label + "' are in the past!");
+				Main.debug("[CalendarEvents] "+ ChatColor.RED+"All events with the label '" + label + "' are in the past!");
 				eventsManager.removeEvent(label);
 				return;
 			}
 		}
-		if(Main.debug)Bukkit.getConsoleSender().sendMessage("next date to schedule: " + ZonedDateTime.ofInstant(Instant.ofEpochMilli(toReturn), zone).toString());
+		Main.debug("next date to schedule: " + ZonedDateTime.ofInstant(Instant.ofEpochMilli(toReturn), zone).toString());
 		nextCall = toReturn;
 	}